Reordering test folders, CONTRIBUTING.md file
