+++ tests/src/stack/TestStack.tsx
@@ -1,6 +1,6 @@
 import { Divider, Stack } from '@chakra-ui/react'
-import { StackTest } from '../../types'
-import TestStatus from '../components/TestStatus'
+import { StackTest } from '@app/types'
+import TestStatus from '@app/components/TestStatus'
 
 const TestStack: StackTest = ({ title, TestStackComponent, run, onComplete, description }) => {
   if (!run) {
