+++ tests/src/stack/UseAuthStack.tsx
@@ -1,44 +1,42 @@
 import { Stack, Divider } from '@chakra-ui/react'
 import TestStatus, { TestStatusType } from '../components/TestStatus'
 import { useEffect, useState } from 'react'
-import { request } from '@lib'
-import { StackComponent } from '../../types'
+import { useAuth } from '@lib'
+import { StackComponent } from '@app/types'
 import { useTestStack } from '@app/contexts/TestStackProvider'
 
-const TEST_STACK_KEY = 'request'
+const TEST_STACK_KEY = 'auth'
 
-const RequestStack: StackComponent = ({ title, onComplete }) => {
+const UseAuthStack: StackComponent = ({ title, onComplete }) => {
   const [testStatus, setTestStatus] = useState<TestStatusType>('pending')
   const { registerNewStack, updateStackFeatures, getStackFeatures } = useTestStack()
+  const [done, setDone] = useState(false)
 
   useEffect(() => {
     registerNewStack(TEST_STACK_KEY)
   }, [])
 
-  useEffect(() => {
-    const fetch = async () => {
-      setTestStatus('running')
+  const auth = useAuth()
 
-      try {
-        updateStackFeatures(TEST_STACK_KEY, { name: 'request', status: 'running' })
-        const res = await request('request-01', { timestamp: Date.now() }, { forceTryAgain: true })
-        updateStackFeatures(TEST_STACK_KEY, { name: 'request', status: 'success', jsonBody: res })
+  useEffect(() => {
+    if (!done) {
+      setTestStatus('running')
+      if (!auth.ready) {
+        updateStackFeatures(TEST_STACK_KEY, { name: 'useAuth', status: 'running', jsonBody: auth })
+      } else {
+        updateStackFeatures(TEST_STACK_KEY, { name: 'useAuth', status: 'success', jsonBody: auth })
         setTestStatus('success')
         onComplete(true)
-      } catch {
-        updateStackFeatures(TEST_STACK_KEY, { name: 'request', status: 'error' })
-        setTestStatus('error')
-        onComplete(false)
+        setDone(true)
       }
     }
-    fetch()
-  }, [])
+  }, [auth, done])
 
   return (
     <Stack mt={4}>
       <Divider />
       <TestStatus
-        test_id="test_request"
+        test_id="test_use_auth"
         title={title}
         status={testStatus}
         features={getStackFeatures(TEST_STACK_KEY)}
@@ -47,4 +45,4 @@ const RequestStack: StackComponent = ({ title, onComplete }) => {
   )
 }
 
-export default RequestStack
+export default UseAuthStack
